--- file_system.py
@@ -1,1 +1,0 @@
-import middle_layer
